Добавил обработчик ошибок для rest-контроллеров.

+++ src/main/java/ru/egarcourses/HospitalInfoSystem/services/impl/CommentaryServiceImpl.java
@@ -8,6 +8,9 @@ import ru.egarcourses.HospitalInfoSystem.models.Commentary;
 import ru.egarcourses.HospitalInfoSystem.repositories.CommentaryRepository;
 import ru.egarcourses.HospitalInfoSystem.services.CommentaryService;
 import ru.egarcourses.HospitalInfoSystem.util.MappingUtils;
+import ru.egarcourses.HospitalInfoSystem.util.exceptions.NotCreatedException;
+import ru.egarcourses.HospitalInfoSystem.util.exceptions.NotFoundedException;
+import ru.egarcourses.HospitalInfoSystem.util.exceptions.NotUpdatedException;
 
 import java.util.List;
 import java.util.Optional;
@@ -28,12 +31,19 @@ public class CommentaryServiceImpl implements CommentaryService {
 
     @Override
     public List<CommentaryDTO> findAll() {
-        return commentaryRepository.findAll().stream().map(mappingUtils::mapToCommentaryDTO).collect(Collectors.toList());
+        List<Commentary> commentaryList = commentaryRepository.findAll();
+        if (commentaryList.isEmpty()) {
+            throw new NotFoundedException("Commentaries not founded");
+        }
+        return commentaryList.stream().map(mappingUtils::mapToCommentaryDTO).collect(Collectors.toList());
     }
 
     @Override
     public CommentaryDTO findById(int id) {
         Optional<Commentary> foundCommentary = commentaryRepository.findById(id);
+        if (!foundCommentary.isPresent()) {
+            throw new NotFoundedException("Commentary not founded");
+        }
         return mappingUtils.mapToCommentaryDTO(foundCommentary.get());
     }
 
@@ -41,6 +51,9 @@ public class CommentaryServiceImpl implements CommentaryService {
     @Override
     public void save(CommentaryDTO commentaryDTO) {
         commentaryRepository.save(mappingUtils.mapToCommentary(commentaryDTO));
+        if (!commentaryRepository.findById(commentaryDTO.getId()).isPresent()) {
+            throw new NotCreatedException("Comment not created");
+        }
     }
 
     @Transactional
@@ -49,11 +62,17 @@ public class CommentaryServiceImpl implements CommentaryService {
         Commentary updatedCommentary = mappingUtils.mapToCommentary(updatedCommentaryDTO);
         updatedCommentary.setId(id);
         commentaryRepository.save(updatedCommentary);
+        if (!commentaryRepository.findById(id).equals(updatedCommentary)) {
+            throw new NotUpdatedException("Comment not updated");
+        }
     }
 
     @Transactional
     @Override
     public void delete(int id) {
+        if (!commentaryRepository.findById(id).isPresent()) {
+            throw new NotFoundedException("Comment not founded");
+        }
         commentaryRepository.deleteById(id);
     }
 }
