+++ src/main/java/ru/egarcourses/HospitalInfoSystem/services/impl/DoctorServiceImpl.java
@@ -8,6 +8,9 @@ import ru.egarcourses.HospitalInfoSystem.models.Doctor;
 import ru.egarcourses.HospitalInfoSystem.repositories.DoctorRepository;
 import ru.egarcourses.HospitalInfoSystem.services.DoctorService;
 import ru.egarcourses.HospitalInfoSystem.util.MappingUtils;
+import ru.egarcourses.HospitalInfoSystem.util.exceptions.NotCreatedException;
+import ru.egarcourses.HospitalInfoSystem.util.exceptions.NotFoundedException;
+import ru.egarcourses.HospitalInfoSystem.util.exceptions.NotUpdatedException;
 
 import java.util.List;
 import java.util.Optional;
@@ -28,12 +31,19 @@ public class DoctorServiceImpl implements DoctorService {
 
     @Override
     public List<DoctorDTO> findAll() {
+        List<Doctor> doctors = doctorRepository.findAll();
+        if (doctors.isEmpty()) {
+            throw new NotFoundedException("Doctors not founded");
+        }
         return doctorRepository.findAll().stream().map(mappingUtils::mapToDoctorDTO).collect(Collectors.toList());
     }
 
     @Override
     public DoctorDTO findById(int id) {
         Optional<Doctor> doctor = doctorRepository.findById(id);
+        if (!doctor.isPresent()) {
+            throw new NotFoundedException("Doctor not found");
+        }
         return mappingUtils.mapToDoctorDTO(doctor.get());
     }
 
@@ -41,6 +51,9 @@ public class DoctorServiceImpl implements DoctorService {
     @Override
     public void save(DoctorDTO doctorDTO) {
         doctorRepository.save(mappingUtils.mapToDoctor(doctorDTO));
+        if (!doctorRepository.findById(doctorDTO.getId()).isPresent()) {
+            throw new NotCreatedException("Doctor not created");
+        }
     }
 
     @Transactional
@@ -49,11 +62,17 @@ public class DoctorServiceImpl implements DoctorService {
         Doctor updatedDoctor = mappingUtils.mapToDoctor(updatedDoctorDTO);
         updatedDoctor.setId(id);
         doctorRepository.save(updatedDoctor);
+        if (doctorRepository.findById(updatedDoctor.getId()).equals(updatedDoctor)) {
+            throw new NotUpdatedException("Doctor not updated");
+        }
     }
 
     @Transactional
     @Override
     public void delete(int id) {
+        if (!doctorRepository.findById(id).isPresent()) {
+            throw new NotFoundedException("Doctor not found");
+        }
         doctorRepository.deleteById(id);
     }
 }
