+++ src/main/java/ru/egarcourses/HospitalInfoSystem/services/impl/RequestServiceImpl.java
@@ -8,6 +8,9 @@ import ru.egarcourses.HospitalInfoSystem.models.Request;
 import ru.egarcourses.HospitalInfoSystem.repositories.RequestRepository;
 import ru.egarcourses.HospitalInfoSystem.services.RequestService;
 import ru.egarcourses.HospitalInfoSystem.util.MappingUtils;
+import ru.egarcourses.HospitalInfoSystem.util.exceptions.NotCreatedException;
+import ru.egarcourses.HospitalInfoSystem.util.exceptions.NotFoundedException;
+import ru.egarcourses.HospitalInfoSystem.util.exceptions.NotUpdatedException;
 
 import java.util.List;
 import java.util.Optional;
@@ -28,12 +31,19 @@ public class RequestServiceImpl implements RequestService {
 
     @Override
     public List<RequestDTO> findAll() {
+        List<Request> requests = requestRepository.findAll();
+        if (requests.isEmpty()) {
+            throw new NotFoundedException("Requests not founded");
+        }
         return requestRepository.findAll().stream().map(mappingUtils::mapToRequestDTO).collect(Collectors.toList());
     }
 
     @Override
     public RequestDTO findById(int id) {
         Optional<Request> foundRequest = requestRepository.findById(id);
+        if (!foundRequest.isPresent()) {
+            throw new NotFoundedException("Request not founded");
+        }
         return mappingUtils.mapToRequestDTO(foundRequest.get());
     }
 
@@ -41,6 +51,9 @@ public class RequestServiceImpl implements RequestService {
     @Override
     public void save(RequestDTO requestDTO) {
         requestRepository.save(mappingUtils.mapToRequest(requestDTO));
+        if (!requestRepository.findById(requestDTO.getId()).isPresent()) {
+            throw new NotCreatedException("Request not created");
+        }
     }
 
     @Transactional
@@ -49,11 +62,17 @@ public class RequestServiceImpl implements RequestService {
         Request updatedRequest = mappingUtils.mapToRequest(updatedRequestDTO);
         updatedRequest.setId(id);
         requestRepository.save(updatedRequest);
+        if (!requestRepository.findById(id).equals(updatedRequest)) {
+            throw new NotUpdatedException("Request not updated");
+        }
     }
 
     @Transactional
     @Override
     public void delete(int id) {
+        if (!requestRepository.findById(id).isPresent()) {
+            throw new NotFoundedException("Request not founded");
+        }
         requestRepository.deleteById(id);
     }
 }
